Move HttpService to service.rs
